clean up views
